src/reducers/todolist-reducer.ts
import {TodoListType} from "../App";
import {v1} from "uuid";

type RemoveTodoListAT = {
    type: "REMOVE-TODOLIST"
    id: string
}

type AddTodoListAT = {
    type: "ADD-TODOLIST"
    title: string
    id: string
}

type ActionType = RemoveTodoListAT | AddTodoListAT

export const todolistsReducer = (todoLists: Array<TodoListType>, action: ActionType): Array<TodoListType> => {
    switch (action.type) {
        case "REMOVE-TODOLIST":
            return todoLists.filter(tl => tl.id !== action.id)
        case "ADD-TODOLIST":
            const newTodoListID = v1()
            return [...todoLists, {id: action.id, title: action.title, filter: 'All'}]
        default:
            return todoLists
    }
}